requirements.txt
google-adk
google-cloud-bigquery
google-generativeai
pandas
python-dotenv
pyarrow
